+++ Demo/Sources/Scenes/Maps/SpecularScene.cpp
@@ -61,5 +61,5 @@ void SpecularScene::Stop() {
 
 void SpecularScene::Update(float sec) {
 	DemoScene::Update(sec);
-	light->GetTransform()->SetPosition(Vector3D(cos(game->GetRunTime() / 2.f)*3.f, 2.f, sin(game->GetRunTime() / 2.f)*3.f));
+	light->GetTransform()->SetPosition(Vector3D(cosf(game->GetRunTime() / 2.f)*3.f, 2.f, sinf(game->GetRunTime() / 2.f)*3.f));
 }
